backend/requirements.txt
fastapi==0.115.8
pydantic==2.10.6
Pillow==11.1.0
ultralytics==8.3.75
uvicorn==0.34.0
torch==2.6.0
numpy==2.1.1

